+++ app/Http/Controllers/ConferenceListController.php
@@ -4,61 +4,103 @@ namespace App\Http\Controllers;
 
 use Illuminate\View\View;
 use App\Models\Conference;
+use App\Models\User;
+use App\Http\Requests\ConferenceRequest;
+use App\Services\RegistrationFormService;
+
 
 use Illuminate\Http\Request;
 use Illuminate\Pagination\Paginator;
-
 use Illuminate\Http\RedirectResponse;
-use App\Http\Requests\RegistrationFormRequest;
 
 
 class ConferenceListController extends Controller
 {
+    /**
+     * Show the conference list with all speakers.
+     */
     public function index(): View
     {
-        //?show description and country 
+        $date = date('Y-m-d');
 
         return view('conference-list', [
-            'conferences' => Conference::with('user')->orderBy('date')->latest()->paginate(10),
+            'conferences' => Conference::where('date', '>=', $date)
+                ->with('user')
+                ->orderByDesc('date')
+                ->paginate(5),
         ]);
     }
 
+    /* [ Functuanility for auth user only ] */
+
     /**
-     * Show the form for creating a new resource.
+     * Save a new created conference to DB.
      */
-    public function create()
+    public function save(ConferenceRequest $request, RegistrationFormService $service)
     {
+        $service->store($request);
+
+        return redirect('/conference_list')->with('success', 'New speaker has been created!');
     }
 
     /**
-     * Display the specified resource.
+     * Show the form for editing the speaker info with all data.
      */
-    public function show()
+    public function edit($id)
     {
-        //
+        $user = User::find($id);
+
+        $countries = ['United Kingdom', 'Poland', 'Germany', 'United States', 'China', 'Japan', 'Ukraine'];
+        array_unshift($countries, $user->country);
+        $countries = array_unique($countries);
+
+        return view('adminPartials/edit', compact('user', 'countries'));
     }
 
     /**
-     * Show the form for editing the specified resource.
+     * Update the specified conference in DB.
      */
-    public function edit()
+    public function update(Request $request, $id)
     {
-        //
+        $user = User::find($id);
+
+        if ($request->file) {
+
+            $photoPath = time() . '.' . $request->file->extension();
+
+            $request->file->move(public_path('userPhotos'), $photoPath);
+
+            $user->photo = $photoPath;
+        }
+
+        $user->firstName = ucfirst($request->input('firstName'));
+        $user->lastName = ucfirst($request->input('lastName'));
+        $user->phone = $request->input('phone');
+        $user->email = $request->input('email');
+        $user->country = $request->input('country');
+
+        if ($user->isDirty()) {
+            $user->save();
+        }
+
+        $user->conferences->title = ucwords($request->input('title'));
+        $user->conferences->description = $request->input('description');
+        $user->conferences->date = $request->input('date');
+
+        if ($user->conferences->isDirty()) {
+            $user->conferences->save();
+        }
+
+        return redirect('/conference_list')->with('success', 'Info has been changed!');
     }
 
     /**
-     * Update the specified resource in storage.
+     * Remove the specified conference from DB.
      */
-    public function update()
+    public function destroy($id)
     {
-        //
-    }
+        /* User::find($id)->delete(); */
 
-    /**
-     * Remove the specified resource from storage.
-     */
-    public function destroy()
-    {
-        //
+        return redirect('/conference_list')->with('success', 'Speaker has been deleted');
     }
 }