+++ resources/views/partials/navigation.blade.php
@@ -13,7 +13,11 @@
 
                         <a href="/conference_list" class="hover:bg-gray-700 hover:text-white px-3 py-2 rounded-md text-base text-white font-medium">Conference List</a>
 
+                        @auth
+                        <a href="/sign_out" class="hover:bg-gray-700 hover:text-white px-3 py-2 rounded-md text-base text-white font-medium">Sign out</a>
+                        @else
                         <a href="/sign_in" class="hover:bg-gray-700 hover:text-white px-3 py-2 rounded-md text-base text-white font-medium">Sign in</a>
+                        @endauth
 
                     </div>
                 </div>