+++ resources/views/sign-in.blade.php
@@ -6,17 +6,18 @@
         <div class="mt-7 bg-white border border-gray-200 rounded-xl shadow-sm">
             <div class="p-4 sm:p-7">
                 <div class="text-center">
-                    <h1 class="block text-2xl text-center font-bold text-gray-800">Sign in</h1>
+                    <h1 class="block text-2xl text-center font-bold text-gray-800">Get Admin Access</h1>
                 </div>
 
                 <div class="mt-5">
-                    <form>
+                    <form method="POST" action="/get_admin_access">
+                        @csrf
                         <div class="grid gap-y-4">
                             <!-- Form Group -->
                             <div>
                                 <label for="email" class="block text-sm mb-2">Email address</label>
                                 <div class="relative">
-                                    <input type="email" id="email" name="email" class="py-3 px-4 block w-full border-gray-200 rounded-md text-sm focus:border-blue-500 focus:ring-blue-500" required aria-describedby="email-error">
+                                    <input type="email" id="email" name="email" value="admin@example.com" class="py-3 px-4 block w-full border-gray-200 rounded-md text-sm focus:border-blue-500 focus:ring-blue-500" required aria-describedby="email-error">
                                     <div class="hidden absolute inset-y-0 right-0 flex items-center pointer-events-none pr-3">
                                         <svg class="h-5 w-5 text-red-500" width="16" height="16" fill="currentColor" viewBox="0 0 16 16" aria-hidden="true">
                                             <path d="M16 8A8 8 0 1 1 0 8a8 8 0 0 1 16 0zM8 4a.905.905 0 0 0-.9.995l.35 3.507a.552.552 0 0 0 1.1 0l.35-3.507A.905.905 0 0 0 8 4zm.002 6a1 1 0 1 0 0 2 1 1 0 0 0 0-2z" />
@@ -33,7 +34,7 @@
                                     <label for="password" class="block text-sm mb-2">Password</label>
                                 </div>
                                 <div class="relative">
-                                    <input type="password" id="password" name="password" class="py-3 px-4 block w-full border-gray-200 rounded-md text-sm focus:border-blue-500 focus:ring-blue-500" required aria-describedby="password-error">
+                                    <input type="password" id="password" name="password" value="adminadmin" class="py-3 px-4 block w-full border-gray-200 rounded-md text-sm focus:border-blue-500 focus:ring-blue-500" required aria-describedby="password-error">
                                     <div class="hidden absolute inset-y-0 right-0 flex items-center pointer-events-none pr-3">
                                         <svg class="h-5 w-5 text-red-500" width="16" height="16" fill="currentColor" viewBox="0 0 16 16" aria-hidden="true">
                                             <path d="M16 8A8 8 0 1 1 0 8a8 8 0 0 1 16 0zM8 4a.905.905 0 0 0-.9.995l.35 3.507a.552.552 0 0 0 1.1 0l.35-3.507A.905.905 0 0 0 8 4zm.002 6a1 1 0 1 0 0 2 1 1 0 0 0 0-2z" />
@@ -45,7 +46,7 @@
                             <!-- End Form Group -->
 
                             <button type="submit" class="py-3 px-4 inline-flex justify-center items-center gap-2 rounded-md border border-transparent font-semibold bg-blue-500 text-white hover:bg-blue-600 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:ring-offset-2 transition-all text-sm">
-                                Sign in
+                                Activate this super mode
                             </button>
                         </div>
                     </form>