+++ routes/web.php
@@ -4,40 +4,35 @@ use Illuminate\Support\Facades\Route;
 
 use App\Http\Controllers\RegistrationFormController;
 use App\Http\Controllers\ConferenceListController;
-use App\Http\Controllers\UserController;
+use App\Http\Controllers\AdminAccessController;
+use App\Models\User;
+
+/*
+ Routes: registration form 
+*/
 
 Route::resource('/', RegistrationFormController::class)
     ->only(['index', 'store']);
 
 Route::view('/edit_form', 'registrationPartials/register');
 
+/*
+ Routes: conference list 
+*/
 Route::get('/conference_list', [ConferenceListController::class, 'index']);
 
-Route::get('/sign_in', [UserController::class, 'signIn']);
-
-/* Route::resource('/list/auth', ListController::class)
-    ->only(['create', 'edit', 'destroy']); */
-/* для авторизованных ->middleware(['auth', 'verified']); */
-
 /*
-|--------------------------------------------------------------------------
-| Web Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register web routes for your application. These
-| routes are loaded by the RouteServiceProvider and all of them will
-| be assigned to the "web" middleware group. Make something great!
-|
+ Routes: admin access
 */
+Route::view('/sign_in', 'sign-in')->middleware('guest');
+Route::view('/sign_out', 'sign-out')->middleware(['auth', 'verified']);
 
-Route::get('/dashboard', function () {
-    return view('dashboard');
-})->middleware(['auth', 'verified'])->name('dashboard');
+Route::post('/get_admin_access', [AdminAccessController::class, 'signIn']);
+Route::post('/leave_admin_access', [AdminAccessController::class, 'signOut']);
 
-Route::middleware('auth')->group(function () {
-    Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
-    Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
-    Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
-});
+Route::view('/add_speaker', 'adminPartials/add');
+Route::get('/edit_speaker/{id}', [ConferenceListController::class, 'edit']);
 
-require __DIR__ . '/auth.php';
+Route::post('/save_new_speaker', [ConferenceListController::class, 'save']);
+Route::post('/update_speaker/{id}', [ConferenceListController::class, 'update']);
+Route::get('/delete_speaker/{id}', [ConferenceListController::class, 'destroy']);