+++ CHANGELOG.md
@@ -5,11 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+- Created 'Sign In' page view with success flash message. 
+- Added DB seeder with admin access ['email' => 'admin@example.com', 'password' => 'adminadmin'].
+- Updated table style in 'Conference List' view and added admin functions.  
+
 # [0.1.1] - 2023-03-15 
 
 ### Fixed 
 - Added date validation with Laravel.
-- Changed PHP syntax to Blade PHP syntax in view('registrationPartials/congratulation').
+- Changed PHP syntax to Blade PHP syntax in 'Congratulation' view.
 
 # [0.1.0] - 2023-03-13